+++ src/app/pages/demo-angular-lib/demo-angular-lib.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { DemoAngularLibRoutingModule } from './demo-angular-lib-routing.module';
+/*-- Import medes-ui --*/ import { MdsHightlightPrismModule } from 'medes-ui';
+// /*-- Medes Team Only --*/ import { MdsHightlightPrismModule } from 'projects/medes-ui/src/public-api';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    FormsModule,
+    DemoAngularLibRoutingModule,
+    MdsHightlightPrismModule
+  ]
+})
+export class DemoAngularLibModule { }