+++ src/app/pages/demo-angular-lib/demo-pipes/demo-pipe-routing.module.ts
@@ -1,14 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { IntroPipeComponent } from './pipe.component';
 import { DemoFilterPipeComponent } from './demo-filter-pipe/demo-filter-pipe.component';
 import { DemoSearchPipeComponent } from './demo-search-pipe/demo-search-pipe.component';
 import { DemoSafeurlPipeComponent } from './demo-safeurl-pipe/demo-safeurl-pipe.component';
 import { DemoFilterRangePipeComponent } from './demo-filter-range-pipe/demo-filter-range-pipe.component';
 
 const routes: Routes = [
-  { path: '', redirectTo: 'intro', pathMatch: 'full'  },
-  { path: 'intro', component: IntroPipeComponent },
+  { path: '', redirectTo: 'filter', pathMatch: 'full'  },
   { path: 'filter', component: DemoFilterPipeComponent },
   { path: 'filter-range', component: DemoFilterRangePipeComponent },
   { path: 'search', component: DemoSearchPipeComponent },