+++ src/app/pages/demo-utils/aray-utils/aray-utils.component.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-/*-- Import medes-ui --*/ import { MdsHightlightPrismModule } from 'medes-ui';
-// /*-- Medes Team Only --*/ import { MdsHightlightPrismModule } from 'projects/medes-ui/src/public-api';
+/*-- Import medes-ui --*/ import { MdsArrayUtils, MdsHightlightPrismModule } from 'medes-ui';
+// /*-- Medes Team Only --*/ import { MdsArrayUtils, MdsHightlightPrismModule } from 'projects/medes-ui/src/public-api';
 
 @Component({
   selector: 'mds-aray-utils',
@@ -20,7 +20,7 @@ MdsArrayUtils.countUniqueValues(["Web","Web","Web","Web","iOS","iOS","Android","
 MdsArrayUtils.sumValues([5,3,4,56,6,7]);
 // output 81
 
-//sample data
+// Sample data
 data = [
   { "manufucture": "audi", "model": "r8", "year": "2012" },
   { "manufucture": "audi", "model": "rs5", "year": "2013" },
@@ -57,7 +57,18 @@ MdsArrayUtils.arrayTolistObject(this.data, manufucture);
   audi: {manufucture: 'audi', model: 'rs5', year: '2013'}
   ford: {manufucture: 'ford', model: 'fusion', year: '2015'}
   kia: {manufucture: 'kia', model: 'optima', year: '2012'}
-}*/`;
+}*/
+
+// Sample data
+const sampleDataExport = [
+  ['name', 'age', 'gender'],
+  ['John', 23, 'male'],
+  ['Doe', 25, 'male'],
+  ['Jesica', 19, 'female']
+]
+
+MdsArrayUtils.exportToCsv('my-file-name', sampleDataExport)
+/* For output click button Download CSV below`;
 
 
 data = [
@@ -68,7 +79,17 @@ data = [
   { manufucture: 'kia', model: 'optima', year: '2012' }
 ];
 
-  // constructor() { }
+  constructor() {}
+
+  downloadCsv(): void {
+    const sampleDataExport = [
+      ['name', 'age', 'gender'],
+      ['John', 23, 'male'],
+      ['Doe', 25, 'male'],
+      ['Jesica', 19, 'female']
+    ]
+    MdsArrayUtils.exportToCsv(sampleDataExport, 'my-file-name');
+  }
 
 
 }