+++ src/app/pages/introduction/how-to-install/how-to-install.component.ts
@@ -6,6 +6,11 @@ import { Component } from '@angular/core';
   styleUrls: ['./how-to-install.component.scss']
 })
 export class HowToInstallComponent {
+  importCss = `
+/*-- MedesUI --*/
+@import \'../node_modules/medes-ui/assets/scss/medes-ui.scss\';
+@import \'../node_modules/medes-ui/assets/scss/ui-colorscheme.scss\';
+// Add your own custom color scheme - https://doc.mediadesain.com/introduction/color-scheme`;
 
   // constructor() { }